+++ lib/home/settings.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:vkhillseva/widgets/loading_overlay.dart';
-import 'package:vkhillseva/common/theme.dart';
-import 'package:vkhillseva/home/festival_settings_page.dart';
+import 'package:vkhgaruda/widgets/loading_overlay.dart';
+import 'package:vkhgaruda/common/theme.dart';
+import 'package:vkhgaruda/home/festival_settings_page.dart';
 
 class Settings extends StatefulWidget {
   final String title;